--- a/cmd/generate/config/rules/generic.go
+++ b/cmd/generate/config/rules/generic.go
@@ -1,7 +1,7 @@
 package rules
 
 import (
-	"regexp"
+	regexp "github.com/wasilibs/go-re2"
 
 	"github.com/zricethezav/gitleaks/v8/cmd/generate/config/utils"
 	"github.com/zricethezav/gitleaks/v8/cmd/generate/secrets"
@@ -127,7 +127,7 @@ func GenericCredential() *config.Rule {
 		`[DEBUG]		org.neo4j.neo4j-graphdb-api:jar:3.5.12:test`,
 		`apiUrl=apigee.corpint.com`,
 		// TODO: Jetbrains IML files (requires line-level allowlist).
-		//`<orderEntry type="library" scope="PROVIDED" name="Maven: org.apache.directory.api:api-asn1-api:1.0.0-M20" level="projcet" />`
+		// `<orderEntry type="library" scope="PROVIDED" name="Maven: org.apache.directory.api:api-asn1-api:1.0.0-M20" level="projcet" />`
 
 		// Auth
 		`author = "james.fake@ymail.com",`,
@@ -156,9 +156,9 @@ func GenericCredential() *config.Rule {
 		`minisat-master-keying:x64-uwp=fail`,
 		`IceSSL.KeyFile=s_rsa1024_priv.pem`,
 		`"bucket_key": "SalesResults-1.2"`,
-		//`<TAR key="REF_ID_923.properties" value="/opts/config/alias/"/>`,
+		// `<TAR key="REF_ID_923.properties" value="/opts/config/alias/"/>`,
 		`<key tag="SecurityIdentifier" name="SecurityIdentifier" type="STRING" />`,
-		//`packageKey":` + newPlausibleSecret(`[a-zA-Z0-9\-_.=]{30}`),
+		// `packageKey":` + newPlausibleSecret(`[a-zA-Z0-9\-_.=]{30}`),
 		`schemaKey = 'DOC_Vector_5_32'`,
 		`sequenceKey = "18"`,
 		`app.keystore.file=env/cert.p12`,
@@ -166,15 +166,15 @@ func GenericCredential() *config.Rule {
 		`	doc.Security.KeySize = PdfEncryptionKeySize.Key128Bit;`,
 		`o.keySelector=n,o.haKey=!1,`,
 		// TODO: Requires line-level allowlists.
-		//`<add key="SchemaTable" value="G:\SchemaTable.xml" />`,
+		// `<add key="SchemaTable" value="G:\SchemaTable.xml" />`,
 		//	`secret:
-		//secretName: app-decryption-secret
-		//items:
+		// secretName: app-decryption-secret
+		// items:
 		//	- key: app-k8s.yml
 		//	  path: app-k8s.yml`,
 
 		// TODO: https://learn.microsoft.com/en-us/windows/apps/design/style/xaml-theme-resources
-		//`<Color x:Key="NormalBrushGradient1">#FFBAE4FF</Color>`,
+		// `<Color x:Key="NormalBrushGradient1">#FFBAE4FF</Color>`,
 
 		// Password
 		`password combination.