--- a/cmd/generate/config/rules/shopify.go
+++ b/cmd/generate/config/rules/shopify.go
@@ -1,7 +1,7 @@
 package rules
 
 import (
-	"regexp"
+	regexp "github.com/wasilibs/go-re2"
 
 	"github.com/zricethezav/gitleaks/v8/cmd/generate/config/utils"
 	"github.com/zricethezav/gitleaks/v8/config"