--- a/config/config.go
+++ b/config/config.go
@@ -3,12 +3,12 @@ package config
 import (
 	_ "embed"
 	"fmt"
-	"regexp"
 	"sort"
 	"strings"
 
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/viper"
+	regexp "github.com/wasilibs/go-re2"
 )
 
 //go:embed gitleaks.toml