### Description:
This PR replaces the standard library regex with [go-re2](https://github.com/wasilibs/go-re2) in all instances except for [newLineRegex](https://github.com/gitleaks/gitleaks/blob/a9e6d8c9a73c848e0e0dcbbc62eed0ebe851be3f/detect/detect.go#L28). The standard library regex engine is more performant than go-re2 for small regex patterns.


### Profile
![Screenshot 2024-12-19 at 10 45 12 AM](https://github.com/user-attachments/assets/574a9998-912d-4ee2-a083-03f9a5ba828c)


### Benchmarks
1. `git log -p` output for this repo. [https://github.com/gitleaks/gitleaks]

Before ~58s:
![Screenshot 2024-12-19 at 10 37 02 AM](https://github.com/user-attachments/assets/e8ae8b03-b60f-4ce1-91c3-3c4b84391a59)

After ~9s:
![Screenshot 2024-12-19 at 11 16 30 AM](https://github.com/user-attachments/assets/6e0b1e11-ac76-4767-8a0c-2f0845b80bd7)

2. gitlab repo [https://gitlab.com/gitlab-org/gitlab]

Before ~49s:
![Screenshot 2024-12-19 at 11 26 41 AM](https://github.com/user-attachments/assets/eecf485d-a20e-47c1-9cf8-1ac3e6eb7240)


After ~14s:
![Screenshot 2024-12-19 at 11 30 07 AM](https://github.com/user-attachments/assets/75c5e844-ae18-4e0d-99ca-2079bd56636e)


3. linux repo [https://github.com/torvalds/linux]

Before ~1m52s:
![Screenshot 2024-12-19 at 11 28 54 AM](https://github.com/user-attachments/assets/3d76b26a-fdeb-4bb1-8d66-fb5050f27166)

After ~1m:
![Screenshot 2024-12-19 at 11 31 45 AM](https://github.com/user-attachments/assets/ddba303e-96ac-4198-98ac-fbe3a7e3eafa)




### Checklist:

* [ ] Does your PR pass tests?
* [ ] Have you written new tests for your changes?
* [ ] Have you lint your code locally prior to submission?