@rgmz 2024-12-19
Is there a memory profile available? When TruffleHog switched, a few people complained about the large jump in memory usage.