@ahrav 2024-12-19
> Is there a memory profile available? When TruffleHog switched, a few people complained about the large jump in memory usage.

I shared one with Zach but can add it here as well. I agree—the cost is noticeable, and the tradeoff might not be worth it. From what I understand, the WASM runtime module has an initial cost of 4MB, which increases with each rule, though the per-rule growth is significantly smaller.

I mentioned this to Zach, but feel free to close if the tradeoff isn't worth it.