@rgmz 2024-12-19
It's a shame that Go doesn't have `#if` directives. Then you could easily toggle between both.