@zricethezav 2024-12-20
christ almighty these numbers are impressive