@ahrav 2024-12-20
> It's a shame that Go doesn't have `#if` directives. Then you could easily toggle between both.

I was reading some C code earlier and thought the exact same thing... 😢