@rgmz 2025-01-19
> It's a shame that Go doesn't have `#if` directives. Then you could easily toggle between both.

Seems it's technically possible: #1706